move mixin back to core, move setting to debug
